renamed files back to normal for store and  app indexes
