+++ src/apps/Toolbox/index.jsx
@@ -1,57 +1,29 @@
 import './App.css';
-import React, { Component } from 'react';
+import React, { Component } from 'react'
+import PropTypes from "prop-types";
+import { compose } from 'recompose'
 import { createMuiTheme } from '@material-ui/core/styles';
 import { ThemeProvider, withStyles } from "@material-ui/styles";
 import { SnackbarProvider, withSnackbar } from 'notistack';
 import { inject, observer } from 'mobx-react'
 import { Provider } from 'mobx-react';
 import MobxStore from '../../shared/stores'
-import Forage from '../../shared/localforage'
-import { BrowserRouter as Router, Link, Route, Switch } from 'react-router-dom'
-
+import { createBrowserHistory } from 'history';
+import { syncHistoryWithStore } from 'mobx-react-router';
+import { Router } from 'react-router';
 import { hot } from 'react-hot-loader'
-import AppShell from './containers/AppShell'
+import { Dashboard } from './views/Dashboard';
+import { Box } from '@material-ui/core';
+import { Shelf } from './views/Shelf'
+import { ShelfButton } from './views/ShelfButton';
+import { Sidebar } from './views/Sidebar';
+import { Scribe } from '../Scribe'
+import { BrowserRouter, Link, Route, Redirect, Switch, withRouter } from 'react-router-dom'
 
 // Initilize Root Store
 const store = new MobxStore()
-const forage = new Forage()
-
-const theme = createMuiTheme({
-    palette: {
-        primary: {
-            light: '#757ce8',
-            main: '#cddc39',
-            dark: '#002884',
-            contrastText: '#fff',
-        },
-        secondary: {
-            light: '#ff7961',
-            main: '#2196f3',
-            dark: '#ba000d',
-            contrastText: '#000',
-        },
-        background: {
-            default: '#FFF'
-            // default: '#313439'
-        }
-    },
-    typography: {
-        useNextVariants: true,
-        // fontSize: 60,
-        // htmlFontSize: 20,
-    },
-});
-
-const styles = {
-    root: {
-        fontWeight: 500,
-        boxShadow: '0 2px 2px rgba(0, 0, 0, 0.25)',
-    },
-    success: { backgroundColor: '#72d67a', color: '#209225' },
-    error: { backgroundColor: '#f3815f', color: '#ad250b' },
-    warning: { backgroundColor: '#ffd475', color: '#b9800f' },
-    info: { backgroundColor: '#6dddff', color: '#1892d0' },
-};
+// const browserHistory = createBrowserHistory()
+// const history = syncHistoryWithStore(browserHistory, store.routing)
 
 const snackbarOptions = {
     maxSnack: 3,
@@ -62,105 +34,115 @@ const snackbarOptions = {
     },
 }
 
-export default withStyles(styles)(
-    (props) =>
-    <Provider forage={forage} store={store} >
-        <ThemeProvider {...{ theme }}>
+const theme = createMuiTheme({
+    palette: {
+        // primary: {
+        //     light: '#757ce8',
+        //     // main: '#cddc39',
+        //     dark: '#002884',
+        //     contrastText: '#fff',
+        // },
+        // secondary: {
+        //     light: '#ff7961',
+        //     // main: '#f44336',
+        //     dark: '#ba000d',
+        //     contrastText: '#000',
+        // },
+        accent: {
+            pink: '#FF0099'
+        },
+        background: {
+            default: '#FFF'
+            // default: '#ede5da'
+            // default: '#313439'
+        }
+    },
+    typography: {
+        fontFamily: "Poppins, 'Helvetica Neue','Helvetica','Arial',sans-serif",
+        useNextVariants: true,
+    },
+});
+
+const styles = theme => ({
+    root: {
+        // color: 'red',
+    },
+})
+
+export default () =>
+    <Provider store={store} >
+        <ThemeProvider theme={theme}>
             <SnackbarProvider {...snackbarOptions}>
-                <RoutedApp />
+                <BrowserRouter>
+                    <Toolbox />
+                </BrowserRouter>
             </SnackbarProvider>
         </ThemeProvider>
     </Provider >
-    )
 
-@withSnackbar
-@inject('forage', 'store')
-@observer
-@hot(module)
-class RoutedApp extends Component {
+const Toolbox = compose(
+    withSnackbar,
+    inject('store'),
+    withStyles(styles),
+    observer,
+    hot(module)
+)(
+    class Toolbox extends Component {
 
-    componentDidMount() {
-        const { servicesStore } = this.props.store
-        const { enqueueSnackbar, closeSnackbar } = this.props
+        componentWillMount() {
+            const { servicesStore } = this.props.store
+            const { enqueueSnackbar, closeSnackbar } = this.props
+            // console.log(this.props)
 
-        // : Pass the calling function for notifications to the store
-        servicesStore.setNotifyFunctions({ enqueueSnackbar, closeSnackbar })
+            // : Pass the calling function for notifications to the store
+            servicesStore.setNotifyFunctions({ enqueueSnackbar, closeSnackbar })
+        }
 
-        // : Initialize the localforage session for videos
-        this.props.forage.videos.init()
+        render() {
+            return (
+                <Box display="flex" flexDirection="row" justifyContent="flex-start" alignItems="flex-start" style={{
+                    position: 'absolute', boxSizing: 'border-box', height: '100%', width: '100%', overflow: 'hidden'
+                }}>
+                    <Route path={`/`} render={(match) => (
+                        <Dashboard
+                            shelf={
+                                <Shelf >
+                                    <ShelfButton tooltip='Scribe' icon='scribe' path='/scribe' color="#f9b54c" active={true} />
+                                </Shelf>
+                            }
+                            sidebar={
+                                <Sidebar>
+
+                                </Sidebar>
+                            }
+                            header={
+                                null
+                            }
+                            currentApp={
+                                    <Route path="/scribe" component={Scribe} />
+                            }
+                        />
+                    )} />
+                    <LinkOverlay />
+                </Box>
+            );
+        }
     }
+)
 
-    render() {
-        // const { services } = this.props.store
-        // const { forage } = this.props
-        // const { videos } = forage
-        // const { loader } = services
-
-        return (
-            <Router>
-                {/* <LinkOverlay /> */}
-                <Route path={`/`} render={(match) => (
-                    <AppShell store={store} match={match} />
-                )} />
-            </Router>
-        );
-    }
+Toolbox.propTypes = {
+    classes: PropTypes.object,
+    closeSnackbar: PropTypes.func,
+    enqueueSnackbar: PropTypes.func,
+    store: PropTypes.object
 }
-
-// const LoadingScreen = observer(() =>
-//   <div style={{
-//     height: '100vh', width: '100vw', background: '#eae1c5', position: 'absolute',
-//     display: 'flex', flexFlow: 'row no-wrap', justifyContent: 'center', alignItems: 'center'
-//   }}>
-//     <h1 style={{ color: '#ec5c5c' }}>...</h1>
-//   </div>
-// )
 
 const LinkOverlay = observer(() =>
     <ul style={{
         position: 'fixed', bottom: 0, zIndex: 9999, right: 20, padding: 10, borderRadius: 4, boxShadow: '0px 1px 5px 0px rgba(0, 0, 0, 0.2), 0px 2px 2px 0px rgba(0, 0, 0, 0.14), 0px 3px 1px -2px rgba(0, 0, 0, 0.12)', listStylePosition: "inside", background: 'lightgrey', paddingTop: 0, paddingBottom: 0, fontSize: 12,
     }}>
         <li><Link to="/">/</Link></li>
-        <li><Link to="/letters">/letters</Link></li>
-        <li><Link to="/letters/publish">/letters/publish</Link></li>
-        <li><Link to="/letters/load">/library/load</Link></li>
-        <li><Link to="/settings">/settings</Link></li>
-        <li><Link to="/splash">/splash</Link></li>
+        <li><Link to="/scribe">/scribe</Link></li>
+        <li><Link to="/scribe/Untitled.docx">/scribe/Untitled</Link></li>
     </ul>
 )
-
-// /** Sample fetch code - MP */
-
-
-// const attendees = document.querySelector("#attendees");
-// console.log('attendees: ', attendees);
-
-// fetch("https://jsonplaceholder.typicode.com/users")
-//     .then(response => response.json())
-//     .then(data => {
-
-//         // Actual data:
-//         console.log('users (external): ', data)
-
-//         // Sample UI code
-//         let html = '';
-//         data.forEach(user => {
-//             html +=
-//                 `<div class="card">
-//                 <h2>${user.name}</h2>
-//                 <div>${user.email}</div>
-//             </div>`
-//         });
-//         attendees.innerHTML = html;
-//     })
-
-// fetch("http://localhost:4000/download?URL=https://www.youtube.com/watch?v=i6USBe9Gr3U")
-//     .then(response => response.json())
-//     .then(data => {
-//         console.log('app -> localhost 4000 data: ', data);
-//     })
-
-
-//     // ytdl(URL, {
-//     //     format: 'mp4'
-//     // }).pipe(res);
