+++ src/index.js
@@ -1,7 +1,6 @@
 import './index.css';
 import React from 'react';
 import ReactDOM from 'react-dom';
-// import Toolbox from './apps/Toolbox';
-import Toolbox from './apps/Toolbox/indexNew'
+import Toolbox from './apps/Toolbox'
 
 ReactDOM.render(<Toolbox />, document.getElementById('root'));
